pyproject.toml
[tool.poetry]
name = "xiuxian-tg-bot"
version = "0.1.0"
description = ""
authors = ["0x77"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "^0.19.0"
httpx = "^0.26.0"
httpcore = "^1.0.2"
python-telegram-bot = "^20.8"
websockets = "^12.0"
requests = "^2.31.0"
pyTelegramBotAPI = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
